This represents a major change in JANA. Here are the more important modifications:
1. jana namespace created. All JANA classes now reside in it and so
must be explicitly referenced via jana:: or "using namespace jana" must
be present.

2. Associated objects added to JObject.

3. Header files cleaned up to remove "using namespace XXX" statements.

4. Renamed dataClassName() with GetDataClassName() to make more consistent
with other accessor methods.

5. Replaced lots of calls to T::className() with calls to T::static_className().

6. Replaced HDCLASSDEF with JOBJECT_PUBLIC macro and moved into JObject.h (it
was in JFactory.h).

7. Replaced old system of printing objects which centered on JFactory_base
classes to one which relies on JObject::toStrings()

8. Removed TestProcessor, TestSource, and TestFactory plugins since these
all also existed in TestSpeed.


Not all of the JANA core library files have been reviewed here. This will
likely be done in the near future.



git-svn-id: https://phys12svn.jlab.org/repos/JANA@318 c5ed4466-e916-0410-8347-b3263e9c103d