Converted to setting BMSDIR variable explicitly in each Makefile where appropriate.
It turns out the MAKEFILE_LIST method implemented in the
last revision does not work on older versions of gmake.
It was turning out to be way too complicated to extract
the BMS directory from the inherent variables so now
all programs, plugins, and libraries define it explicitly
relative to their location. This should be much more
robust.


git-svn-id: https://phys12svn.jlab.org/repos/JANA@338 c5ed4466-e916-0410-8347-b3263e9c103d