Moved several files and directories up to be parallel with src. This includes most everything except actual source code files and directories.
The primary motivation was to move the build system a little closer
to the classic unix style where one builds with the recipe
./configure
make
make install

The "make" stage builds everything in the local working directory
while "make install" copies the binaries and headers to the final
installation directory (optionally specified with --prefix=path
when running configure). This allows one to build in a private
user account, but run "make install" using an admin account for
the final installation.

Also included here is a modification to the xerces m4 macro used
by configure to look for xerces in the directory tree specified
by the XERCESCROOT environment variable.


git-svn-id: https://phys12svn.jlab.org/repos/JANA@706 c5ed4466-e916-0410-8347-b3263e9c103d