Fixed bug where factories added from plugins did NOT get preference.
The advertised behavior had been that when a plugin provided a factory
that produced the same data type as one statically compiled into the
executable, the one from the plugin would be used. Prior to now, this
was not the case. One still needs to be cautious about using the same
names for the factory and factory generator classes as is used in the
statically linked executable. Similarly named classes can be linked to
the methods in the executable rather than the plugin giving unexpected
behavior.


git-svn-id: https://phys12svn.jlab.org/repos/JANA@1011 c5ed4466-e916-0410-8347-b3263e9c103d