Fixed bug in JParameterManager where the first thread to call SetDefaultParameter could end up with a different value than subsequent threads. This happened because the first call would store the value as a string, but then return without touching the variable used to pass it in. Subsequent threads would copy in a value what was converted to and from a string ans so could be slightly different.
In addition, values like NaN do not convert back into NaN after storing as a string.
Because of this, a check is now made to verify that the value obtained after
converting to/from a string exactly matches the input value. If this is not the
case then a warning message is printed.


git-svn-id: https://phys12svn.jlab.org/repos/JANA@1244 c5ed4466-e916-0410-8347-b3263e9c103d