Reverted form of virtual method Init() to it's original which does not include any arugments. This is because the DApplication subclass signature did not match the new form and so was not getting called.
The functionality of the boolean argument was moved to a member
valriable create_event_buffer_thread that can be cleared before
Init is called to get the same effect. jcalibread will be updated
to do this on the next commit.

There were also several lines of JApplication.h that were adjusted
for cosmetic purposes.


git-svn-id: https://phys12svn.jlab.org/repos/JANA@1255 c5ed4466-e916-0410-8347-b3263e9c103d