Bugfix: Topology now stops clock as soon as last sink finishes
Topology is now an Activable which is subscribed to its sinks, who notify it when they finish so it can stop the clock and change run state as soon as possible. TopologyBuilder now modifies an existing Topology in-place, as opposed to returning a copy. This was needed because the sinks now keep a pointer to the Topology.