Experiment: Better type safety and efficiency when handling JFactory*s
The basic idea is to change a bunch of return types on templated functions from JFactory* to  JFactoryT<T>*. Also added a templated GetFactory<T>. This gets rid of a lot of virtual method calls to JFactoryT::Set. It also eliminates repackaging of vectors of pointers in order to cast vector<T*> => vector<JObject*> and vice versa.