Skip compilation of targets missing dependencies in CMake
The overall goal is to bring CMake builds up to parity with SCons builds.

First and foremost we have to handle missing dependencies such as ROOT and ZMQ. The desired behavior is that the build system would detect that these libraries are missing, offer the user a hint about how to find them, and proceed with compiling everything else.

This commit makes CMake target generation conditioned on the packages being found. There are many improvements possible; this is just a first step.