Inform user of cmake install directory and reason during install stage
Previously, the cmake build script informed the user where they were installing to and why, but not during the 'install' stage. Since the CMake project gets reloaded far less often than it gets recompiled, the message is invisible at the point when it is needed. This change ensures the message gets printed after the user selects "Install" in CLion or "make install" on the command line instead.