Generate jana_config script from CMake
The cmake build script will now call out to a python script to generate a bash script which gets installed to $JANA_HOME/bin. The python script is a stub for now because we still have to extract it from the SBMS code. Naturally, it in turn calls a Perl script which produces the OSNAME string. Perhaps we could skip all of this indirection and just use a CMake jinja template.

This took some research because CMake makes the relationship between files and targets confusing when you stray off of the happy path. The trick is to use a dummy custom target as the glue between a custom command and an install.