Install public header files from CMake
The CMake build script will now place JANA's public header files under $JANA_HOME/include. This is a key detail allowing us to create plugins which live outside the jana2 repository. The question "what is a public header?" remains open. Currently, all headers are effectively public, although only the top-level headers are intended to be stable. We may find it beneficial to refactor such that we only expose the top-level headers to the user, and maybe separate out the public headers entirely.