Generate jana_config.h via CMake
This commit adds support for generating the jana_config.h header through CMake, which is necessary for replacing SCons when the time comes. It also now installs header files to include/JANA/ rather than include/, which was an earlier oversight.

Further work is needed to verify that XERCES, CCDB, etc are correctly located when built on different systems. Much of the logic in make_config.cmake ought to be refactored into Find*.cmake files. In general there is confusion between finding libraries installed on the system vs in a user directory vs on a network share, and those files are the place to isolate that complexity from the rest of the build system.