Add test cases to verify JFactory behavior
Work on porting BDX revealed that some JFactory behavior was not fully
implemented or working as intended. In particular, this involves
persistent JFactories, non-object-owner JFactories, JFactories populated
from JEventSource::GetObjects, and calling ChangeRun() at the correct time.

Complicating things, the JFactory implementation on master has diverged
due to other bugfixes, not all of which are compatible with the changes
for BDX. In order to bring all of the BDX changes onto master as well,
we need to precisely express the desired behavior of JFactories and verify
that this behavior has been correctly implemented. These test cases are
a first attempt at this.