Expose JEventSourceBackend's event count, status, etc
This commit will expose all of the bookkeeping associated with a JEventSource which is overtly NOT the user's responsibility, such as event count, status, plugin name, etc, and extend the test cases to verify that the status and event count are doing the correct thing along the happy path. The information is used by JComponentManager and JStatus.