Apply patch of changes arising from the BdxReco port
The commit history graph for bdx_refactoring is very confusing at this
point. Some changes are needed for tridas_integration, whereas others
introduce new problems we don't really want to deal with right now.
The goal is to separate these two concerns so we can get what we need into
master.

This commit establishes the baseline set of bdxReco-specific modifications
as a single update to the tip of master. Next we will revert any changes
which cause us unnecessary headaches (e.g. to jout) and ensure that the
tests (JFactoryTests in particular) once again all pass.

This partly addresses issue #42.