Ensure JEventProcessors are always furnished with a JApplication
If the user doesn't pass the JApplication pointer via a constructor,
the JComponentManager will step in and set it before anyone can
call JEventProcessor::Init.

This removes one cause of segfaults, and reduces the amount of
boilerplate needed to create a JEventSource. The user no longer needs to
create a throwaway constructor for their JEventSource simply to pass
the JApplication pointer to the parent constructor. It also supports
the broader pattern of moving logic out of the constructor and into
Init.