Add metadata functionality to JFactoryT, JEvent
Oftentimes a user may want to access meta-information about JFactory::Process. For instance, supposing a JFactory<Track> produces a vector of Track objects given a vector of TrackCandidates, the user probably also wants to know statistics about the TrackCandidates that failed to fit at all and hence didn't produce any Tracks. Previously, they would define member variables on the JFactory itself to store this data, access it as a parent  JFactoryT<T>* via GetFactory<T>, and casting it back to the child class e.g. JFactory_GenFitTrackFitter. This breaks the principle of Liskov substitutability and weakens the reusability of our plugin architecture.

JMetadata provides the metadata functionality without these drawbacks. It defines a struct JMetadata<T>, which the user may specialize for their particular JObject class. For instance, JMetadata<Track> could contain a failed_track_count. Any JFactoryT<Track> is able to furnish the user with a JMetadata<Track>, not just JFactory_GenFitTrackFitter. The user should call JEvent::GetMetadata<T>("tag") to access it from another component, and should call JFactory::SetMetadata() from inside JFactory::Process to access it. The metadata object is tied to the lifetime of the JFactory and does *not* get cleared when the JFactory is recycled (unlike the data), so the user should be careful to reset the appropriate fields whenever Process() is called.