Cherry picked changes from pull request #45.
This commit adds JEvent::GetAll(), allowing the user to access the JFactories and JObjects of some type T for all tags, as opposed to for a single tag. This allows the user to access the results of factories without knowing the names of the factory tags a priori, improving the decoupling between plugins.

This commit partially addresses issue #42.