Improve user-friendliness of JPluginLoader
Previously, when a plugin failed to load, JPluginLoader would only show debugging information if jana:debug_plugin_loading=1. Now, the most useful information, namely the list of paths checked and a brief status of what happened when JPluginLoader examined that path, is included in the error message all the time. This saves everybody a lot of time.