Update ccpp.yml
For some reason, cmake on the github system is not passing CMAKE_INSTALL_PREFIX to the cache file or cmake_install.cmake file. This in spite of our output when cmake is running clearly indicating the directory we specify via -DCMAKE_INSTALL_PREFIX as the one where files will be installed. This causes the make install to fail since it tries writing to the root directory / . Disabling the actual "make install" invocation for now.