Fix memory leaks
Second round of memory leak fixes. This time we give ownership of event source specific JFactoryGenerators to JComponentManager. Note that event source specific JFactoryGenerators don't do quite what we want them to, so we will probably go back and rework this. In the future, we may want them to be owned by the JEventSource instead.

This commit addresses issue #55.