Fix potential use-after-free error
Components are owned by JComponentManager yet accessed by JArrowProcessingController and JArrowTopology. If the ServiceLocator destroys JComponentManager first, it will use-after-free any JEventProcessors when attempting to destroy the JArrowProcessingController. Luckily, because we are using shared_ptr for services, we can ensure that the lifetime of JComponentManager always exceeds that of JArrowProcessingController and JTopology by hanging on to a copy of the shared_ptr. This bug never manifested - it was found by AddressSanitizer instead.