Improve build system user-friendliness
Previously, we required users to explicitly provide CMAKE_INSTALL_PREFIX.
This is more strict than necessary. Now we default it to somewhere in
the build directory instead. This achieves the previously stated goals
of not silently clobbering some shared install directory, and it is more
user-friendly in the sense that they can verify that the build works without
committing to a permanent install directory ahead of time.