Improve user-friendliness of jana-generate.py
Some generators require multiple arguments, and it was not clear from the UI which arguments were needed when. If the incorrect arguments were specified, it would print a Python stack trace. Now, it prints a friendlier error message explaining what each generator does and which arguments it needs.

The commit partially addresses issue #64.