Control whether FreeEvent gets called via a flag
FreeEvent locks a mutex, which introduces contention between the JES arrow and the JEP arrow. Since we don't really want users to be doing this anyhow, we make this opt-in. I strongly believe that
uses of FreeEvent should be replaced by JEventGroups or shared_ptrs to JObjects.

Note we had other options:
1. Always call FreeEvent, using a flag to control whether or not we lock the mutex. Encourage users to only write to atomic flags inside FreeEvent so that the mutex is not even necessary. Rejected because it assumes too much user knowledge about multithreading and makes FreeEvent seem more permanent.

2. Only call FreeEvent (always with mutex) when flag is set. Chosen because it is lightweight, safe, and communicates to the user that FreeEvent is discouraged.

3. Maintain a queue of events waiting to be freed. Call FreeEvent from inside the JESA. Rejected because it is technically complicated, has unclear performance implications, and makes FreeEvent seem more permanent.