Rename 'FreeEvent' to 'FinishEvent'
'FreeEvent' is a dangerous misnomer because it encourages users to free resources attached to the event.
They should never do this. Instead, they should use this to notify the EventSource that an event has completely
finished processing. This likely involves freeing resources attached to the JEventSource. The method has been renamed
to reflect its proper use.