Add JFactory::SetTag
This is a step towards solving a longstanding headache with tagged JFactories. Previously, the user was supposed to provide both the object name and the tag as constructor arguments. The object name was defaulted to the demangled T, and the tag was defaulted to the empty string. However, there was no circumstance where we would want the user to choose a _different_ object name, no elegant way of defaulting the name, and of course the tag was the _second_ default argument. Effectively, the user was forced to learn all about JTypeInfo::demangle in order to simply set a tag. If they set aName to any other string it would cause confusion and possibly bugs which wouldn't be detected for a long time. Because the constructor took two strings with defaults, the argument order couldn't be reversed and the 'aName' argument couldn't be removed without silently breaking all existing code.

The plan is to deprecate the old JFactoryT constructor in favor of the zero-argument constructor, and just use SetTag. This will also help with the 'tagged factory generators' feature request.