JParameter is a "data object" again
Previously I wanted JParameter to be an implementation detail of JParameterManager and not let its implementation leak into user code, but that appears unrealistic now. So I made it a class again and brought back normal getters and setters. At least this is consistent with everything else and not going to confuse anybody.