Rename 'stop' to 'pause'
This clears up some deep confusion about whether finish() has been called afterwards. In a nutshell: pause=arrows are not running, but have not been finalized, so that can be resumed. stop=arrows are not running and cannot be resumed: they are finished. Stop makes sense for the topology as a whole but doesn't make sense for arrows.