JCM: Suppress exceptions coming from JFactory::Init
The JComponentManager runs JFactory::Init() for each factory in the set, simply so it can collect JParameters before processing. Sometimes Init() throws an exception (e.g. due to missing geometry). If the factory was never going to be called anyway, that exception shouldn't be thrown anyhow, and if the factory does gets called, the exception will be thrown then. So there's no harm in suppressing exceptions here.