Prototype replacement JLogMessage
JLogMessage flushes its output when it receives an End{} token. This is the worst of all worlds because it
1. Prevents us from appending metadata on each line of a multiline message like JStreamLog
2. Prevents us from being able to use it with output-agnostic methods like `void PrintUsefulOutput(std::ostream&)`
3. Ties us to always using C++ stream operators
4. Doesn't let us use C++ stream manipulators
5. Can't be used as a replacement for `jout` and `_DBG_`, which use std::endl and _don't_ have an end-of-message token

The replacement extends `std::stringstream` instead of wrapping it. It flushes on destruction, and users may optionally trigger a flush by streaming in a special flush token. This solves the problems listed above and should be sufficient to replace the existing JLogMessage, jout, jerr, _DBG_, and JStreamLog.