Remove JLogMessage::Flush
This is for the sake of having exactly one way of doing things. Flush complicates the semantics and implementation of JLogMessage without adding any needed functionality. Furthermore it is conceptually confused since a std::stringstream already has a flush() method which does not do what Flush() does. The pared-down JLogMessage is easy to explain as "std::ostringstream with a logger-specific line prefix and RAII for printing its contents to the log".