JException has a JBacktrace instead of a string
Previously, JException would stringify the JBacktrace in the constructor. This is fine as long as JException represents an unrecoverable error. If used for flow control, however, this imposes a steep performance penalty. The penalty will become even steeper when we enable `addr2line`. Now the JException only stringifies the backtrace upon printing, and only if the backtrace is requested.