Put logger mutex behind a preprocessor flag
The logger mutex is not necessary on the platforms we support. Omitting it causes lots and lots of false positives when running tools like Helgrind. However, including it causes some extremely weird behavior when Helgrind runs, such as segfaults and hanging. I speculate that this behavior is due to the ODR violation that happens with the static logger mutex when JANA is linked incorrectly