Remove untyped wiring from JOmniFactoryGenerator
Even though it looks very similar to the typed wiring, it behaves unexpectedly differently when the target OmniFactory uses Parameters that are not members of the ConfigT object. (Specifically, those parameters are silently dropped.)

JOmniFactory now always uses ConfigT and typed wiring. JWiringService always uses untyped wiring, and it works with any Component whether there is a ConfigT or not.