Event is always cleared upon returning to pool
Previously it wasn't cleared when the event was being returned upstream.
This was meant to save on performance when waiting on a slow socket.
However, it needs to be cleared when nskip is used, or in any other context. While
this could have been done manually on the JEventSourceArrow, the approach I chose
is conceptually a lot simpler because it means that the lifetime of the
data contained inside the JEvent is constrained to the event's in-flight time.