Simplify JBenchUtils::consume_cpu_ms() behavior
Previously, it would either perform a variable amount of work for a
fixed amount of time, or perform a fixed amount of work. The latter
option was not representative of real reconstruction computation, nor
was did the input argument have units of FLOPS, nor was it apparent from
the call site which strategy was intended. JBenchUtils is presently used
exclusively for sanity-checking performance by introducing predictable
bottlenecks, i.e. the former case.