Add JEventSource::Skip
This allows JANA2 to skip all the way to the desired event, if possible, without having the overhead of repeatedly firing the JEventSourceArrow and popping from and pushing to the event pool. When overridden by the user, it enables traversing the input file without having to read the skipped events at all.

This addresses issue #422.